Fix client undefined on project edit function

+++ app/Http/Controllers/ProjectController.php
@@ -291,12 +291,14 @@ class ProjectController extends Controller
         $project->started_at = request()->start_at;
         $project->end_at = request()->end_at;
 
-        if(request()->has('client_id')){
+        $client_old = $project->getClient();
+
+        if(request()->has('client_id') && $client_old->id != request()->client_id){
             if(count($project->client) == 0) {
-                $project->members()->attach(request()->client_id, ['role' => 'client']);
+                $project->members()->attach(request()->client_id, ['role' => 'Client']);
             } else if (isset($project->client()->first()->id) && $project->client()->first()->id != request()->client_id) {
                 $project->members()->detach($project->client()->first()->id);
-                $project->members()->attach(request()->client_id, ['role' => 'client']);
+                $project->members()->attach(request()->client_id, ['role' => 'Client']);
             }
         }
